requirements.txt
awscli==1.37.26
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
docutils==0.16
Flask==3.1.0
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
lxml==5.3.1
MarkupSafe==3.0.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pip==25.0.1
pyasn1==0.6.1
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
python-docx==1.1.2
awscli             1.37.26
awsebcli           3.21.0
blinker            1.9.0
boto3              1.36.26
botocore           1.35.99
cement             2.10.14
certifi            2025.1.31
cffi               1.17.1
charset-normalizer 3.4.1
click              8.1.8
colorama           0.4.6
cryptography       44.0.1
docutils           0.16
Flask              3.1.0
fpdf               1.7.2
idna               3.10
itsdangerous       2.2.0
Jinja2             3.1.5
jmespath           1.0.1
lxml               5.3.1
MarkupSafe         3.0.2
pathspec           0.10.1
pdfminer.six       20231228
pdfplumber         0.11.5
pillow             11.1.0
pip                25.0.1
pyasn1             0.6.1
pycparser          2.22
pypdfium2          4.30.1
pypiwin32          223
python-dateutil    2.9.0.post0
python-docx        1.1.2
pywin32            308
PyYAML             6.0.2
requests           2.32.3
rsa                4.7.2
s3transfer         0.11.2
semantic-version   2.10.0
setuptools         75.8.0
six                1.17.0
termcolor          2.5.0
typing_extensions  4.12.2
urllib3            1.26.20
wcwidth            0.2.13
Werkzeug           3.1.3